Final commit

+++ app/Http/Controllers/SparepartController.php
@@ -6,7 +6,9 @@ use App\Models\Sparepart;
 use App\Models\Kategori;
 use App\Models\Brand;
 use App\Models\Unit;
+use App\Models\Supplier;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule; //nama, ktegri, brnd, unit unik
 
 class SparepartController extends Controller
 {
@@ -21,42 +23,45 @@ class SparepartController extends Controller
         $spareparts = Sparepart::with([
             'kategori',
             'brand',
-            'unit'
+            'unit',
+            'suppliers',
         ])
-
             ->when($search, function ($query) use ($search) {
-
                 $query->where(function ($q) use ($search) {
 
                     $q->where('kode_sparepart', 'like', "%{$search}%")
-                        ->orWhere('nama_sparepart', 'like', "%{$search}%");
+                        ->orWhere('nama_sparepart', 'like', "%{$search}%")
+
+                        ->orWhereHas('suppliers', function ($supplier) use ($search) {
+                            $supplier->where('kode_supplier', 'like', "%{$search}%")
+                                ->orWhere('nama_supplier', 'like', "%{$search}%")
+                                ->orWhere('alamat', 'like', "%{$search}%")
+                                ->orWhere('notlp', 'like', "%{$search}%")
+                                ->orWhere('email', 'like', "%{$search}%")
+                                ->orWhere('stok', 'like', "%{$search}%");
+                        });
                 });
             })
 
             ->when($status, function ($query) use ($status) {
 
-                if ($status == 'low_stock') {
+                if ($status == 'out_stock') {
+                    $query->where('stok', 0);
+                }
 
-                    $query->whereColumn(
-                        'stok',
-                        '<=',
-                        'min_stok'
-                    );
+                if ($status == 'low_stock') {
+                    $query->where('stok', '>', 0)
+                        ->whereColumn('stok', '<=', 'min_stok');
                 }
 
                 if ($status == 'safe_stock') {
-
-                    $query->whereColumn(
-                        'stok',
-                        '>',
-                        'min_stok'
-                    );
+                    $query->whereColumn('stok', '>', 'min_stok');
                 }
             })
             ->latest()
             ->paginate(10)
             ->withQueryString();
-            
+
         $kodeSparepart = Sparepart::generateKode();
 
         $kategori = Kategori::where(
@@ -100,6 +105,13 @@ class SparepartController extends Controller
     {
         $kodeSparepart = Sparepart::generateKode();
 
+        $suppliers = Supplier::where(
+            'status_supplier',
+            'aktif'
+        )
+            ->orderBy('nama_supplier')
+            ->get();
+
         $kategori = Kategori::where(
             'status_kategori',
             'aktif'
@@ -127,7 +139,8 @@ class SparepartController extends Controller
                 'kodeSparepart',
                 'kategori',
                 'brand',
-                'unit'
+                'unit',
+                'suppliers'
             )
         );
     }
@@ -138,16 +151,10 @@ class SparepartController extends Controller
             'kategori',
             'brand',
             'unit',
-            'suppliers'
+            'suppliers',
+            'transaksiDetails.transaksi.user',
+            'transaksiDetails.transaksi.supplier',
         ]);
-
-        // $sparepart->load([
-        //     'kategori',
-        //     'brand',
-        //     'unit',
-        //     'suppliers',
-        //     'stokTransaksi'
-        // ]);
 
         return view(
             'spareparts.show',
@@ -155,53 +162,123 @@ class SparepartController extends Controller
         );
     }
 
+    /**
+     * Detail Sparepart dari Trash
+     */
+    public function showTrash($id)
+    {
+        $sparepart = Sparepart::withTrashed()
+            ->with([
+                'kategori',
+                'brand',
+                'unit',
+                'suppliers',
+                'transaksiDetails.transaksi.user',
+                'transaksiDetails.transaksi.supplier',
+            ])
+            ->findOrFail($id);
+
+
+        return view(
+            'spareparts.trash-show',
+            compact('sparepart')
+        );
+    }
+
 
     public function store(Request $request)
     {
-        $data = $request->validate([
+        $data = $request->validate(
+            [
+                'nama_sparepart' => [
+                    'required',
+                    'max:100',
+                    Rule::unique('spareparts')
+                        ->where(function ($query) use ($request) {
+                            return $query
+                                ->where('kategori_id', $request->kategori_id)
+                                ->where('brand_id', $request->brand_id)
+                                ->where('unit_id', $request->unit_id);
+                        }),
+                ],
 
-            'nama_sparepart' => 'required|max:100',
+                'kategori_id' => 'required|exists:kategoris,id',
+                'brand_id' => 'required|exists:brands,id',
+                'unit_id' => 'required|exists:units,id',
+                'min_stok' => 'required|integer|min:0',
+                'deskripsi' => 'nullable|string',
 
-            'kategori_id' => 'required|exists:kategoris,id',
+                'suppliers' => 'required|array|min:1',
+                'suppliers.*.supplier_id' => 'required|exists:suppliers,id',
+                'suppliers.*.harga_beli' => 'required|numeric|min:0',
+            ],
 
-            'brand_id' => 'required|exists:brands,id',
-
-            'unit_id' => 'required|exists:units,id',
-
-            'min_stok' => 'required|integer|min:0',
-
-            'deskripsi' => 'nullable|string'
-
-        ]);
+            [
+                'nama_sparepart.unique' =>
+                'Sparepart dengan nama, kategori, brand, dan unit yang sama sudah terdaftar.',
+            ]
+        );
 
         $data['kode_sparepart'] = Sparepart::generateKode();
 
-        /*
-            STOK SELALU DIMULAI DARI 0
-
-            nanti bertambah
-            melalui Barang Masuk
-        */
-
+        // stok awal selalu 0
         $data['stok'] = 0;
 
-        Sparepart::create($data);
+        $supplierIds = collect($request->suppliers)
+            ->pluck('supplier_id');
+
+        if ($supplierIds->count() !== $supplierIds->unique()->count()) {
+
+            return back()
+                ->withInput()
+                ->withErrors([
+                    'suppliers' => 'Supplier tidak boleh dipilih lebih dari satu kali.'
+                ]);
+        }
+
+        // simpan sparepart
+        $sparepart = Sparepart::create([
+
+            'kode_sparepart' => $data['kode_sparepart'],
+            'nama_sparepart' => $data['nama_sparepart'],
+            'kategori_id' => $data['kategori_id'],
+            'brand_id' => $data['brand_id'],
+            'unit_id' => $data['unit_id'],
+            'min_stok' => $data['min_stok'],
+            'stok' => $data['stok'],
+            'deskripsi' => $data['deskripsi'],
+        ]);
+        //  Simpan relasi ke tbl pivor
+
+        $pivotData = [];
+
+        foreach ($request->suppliers as $supplier) {
+            $pivotData[$supplier['supplier_id']] = [
+                'harga_beli' => $supplier['harga_beli']
+            ];
+        }
+
+        $sparepart->suppliers()->sync($pivotData);
 
         return redirect()
-
             ->route('spareparts.index')
-
             ->with(
                 'success',
                 'Sparepart berhasil ditambahkan.'
             );
     }
 
-    /**
-     * Form edit
-     */
     public function edit(Sparepart $sparepart)
     {
+        $kodeSparepart = Sparepart::generateKode();
+
+        $suppliers = Supplier::where(
+            'status_supplier',
+            'aktif'
+        )
+            ->orderBy('nama_supplier')
+            ->get();
+
         $kategori = Kategori::where(
             'status_kategori',
             'aktif'
@@ -229,43 +306,83 @@ class SparepartController extends Controller
                 'sparepart',
                 'kategori',
                 'brand',
-                'unit'
+                'unit',
+                'kodeSparepart',
+                'suppliers'   // <-- WAJIB ditambahkan
             )
         );
     }
-
     /**
      * Update
      */
     public function update(Request $request, Sparepart $sparepart)
     {
-        $data = $request->validate([
+        $data = $request->validate(
+            [
+                'nama_sparepart' => [
+                    'required',
+                    'max:100',
+                    Rule::unique('spareparts')
+                        ->ignore($sparepart->id)
+                        ->where(function ($query) use ($request) {
+                            return $query
+                                ->where('kategori_id', $request->kategori_id)
+                                ->where('brand_id', $request->brand_id)
+                                ->where('unit_id', $request->unit_id);
+                        }),
+                ],
+                'kategori_id' => 'required|exists:kategoris,id',
+                'brand_id' => 'required|exists:brands,id',
+                'unit_id' => 'required|exists:units,id',
+                'min_stok' => 'required|integer|min:0',
+                'deskripsi' => 'nullable|string',
 
-            'nama_sparepart' => 'required|max:100',
+                'suppliers' => 'required|array|min:1',
+                'suppliers.*.supplier_id' => 'required|exists:suppliers,id',
+                'suppliers.*.harga_beli' => 'required|numeric|min:0',
+            ],
 
-            'kategori_id' => 'required|exists:kategoris,id',
+            [
+                'nama_sparepart.unique' =>
+                'Sparepart dengan nama, kategori, brand, dan unit yang sama sudah terdaftar.',
+            ]
+        );
 
-            'brand_id' => 'required|exists:brands,id',
+        /* g boleh update stok dari halaman master. */
 
-            'unit_id' => 'required|exists:units,id',
+        $supplierIds = collect($request->suppliers)
+            ->pluck('supplier_id');
 
-            'min_stok' => 'required|integer|min:0',
+        if ($supplierIds->count() !== $supplierIds->unique()->count()) {
 
-            'deskripsi' => 'nullable|string'
+            return back()
+                ->withInput()
+                ->withErrors([
+                    'suppliers' => 'Supplier tidak boleh dipilih lebih dari satu kali.'
+                ]);
+        }
 
+        $sparepart->update([
+            'nama_sparepart' => $data['nama_sparepart'],
+            'kategori_id'    => $data['kategori_id'],
+            'brand_id'       => $data['brand_id'],
+            'unit_id'        => $data['unit_id'],
+            'min_stok'       => $data['min_stok'],
+            'deskripsi'      => $data['deskripsi'],
         ]);
 
-        /*
-            Tidak boleh update stok
-            dari halaman master.
-        */
+        $pivotData = [];
 
-        $sparepart->update($data);
+        foreach ($request->suppliers as $supplier) {
+            $pivotData[$supplier['supplier_id']] = [
+                'harga_beli' => $supplier['harga_beli']
+            ];
+        }
+
+        $sparepart->suppliers()->sync($pivotData);
 
         return redirect()
-
             ->route('spareparts.index')
-
             ->with(
                 'success',
                 'Sparepart berhasil diperbarui.'
@@ -277,6 +394,14 @@ class SparepartController extends Controller
      */
     public function destroy(Sparepart $sparepart)
     {
+        if ($sparepart->transaksiDetails()->exists()) {
+
+            return back()->with(
+                'error',
+                'Sparepart memiliki riwayat transaksi sehingga tidak dapat dihapus.'
+            );
+        }
+
         $sparepart->delete();
 
         return back()->with(
@@ -353,7 +478,7 @@ class SparepartController extends Controller
             );
         }
 
-        if ($sparepart->stokTransaksi()->exists()) {
+        if ($sparepart->transaksiDetails()->exists()) {
 
             return back()->with(
                 'error',
